backend/requirements.txt
fastapi
uvicorn
pypdf
Python-docx
psycopg2
APScheduler
bs4
requests 
python-dotenv